enable Rubocop

+++ lib/autobuild/package.rb
@@ -1,10 +1,10 @@
 module Autobuild
-    TARGETS = %w{import prepare build}
+    TARGETS = %w[import prepare build].freeze
 
     class << self
         attr_accessor :ignore_errors
     end
-    
+
     # Basic block for the autobuilder
     #
     # The build is done in three phases:
@@ -43,14 +43,18 @@ module Autobuild
         # The set of utilities attached to this package
         # @return [{String=>Utility}]
         attr_reader :utilities
+
         # Whether {#apply_post_install} has been called
-        def applied_post_install?; !!@applied_post_install end
-        
+        def applied_post_install?
+            @applied_post_install
+        end
+
         # Sets importer object for this package. Defined for backwards compatibility.
         # Use the #importer attribute instead
         def import=(value)
             @importer = value
         end
+
         # Sets an importer object for this package
         attr_accessor :importer
 
@@ -74,11 +78,20 @@ module Autobuild
         end
 
         # Absolute path to the source directory. See #srcdir=
-        def srcdir; File.expand_path(@srcdir || name, Autobuild.srcdir) end
+        def srcdir
+            File.expand_path(@srcdir || name, Autobuild.srcdir)
+        end
+
         # Absolute path to the import directory. See #importdir=
-        def importdir; File.expand_path(@importdir || srcdir, Autobuild.srcdir) end
+        def importdir
+            File.expand_path(@importdir || srcdir, Autobuild.srcdir)
+        end
+
         # Absolute path to the installation directory. See #prefix=
-        def prefix; File.expand_path(@prefix || '', Autobuild.prefix) end
+        def prefix
+            File.expand_path(@prefix || '', Autobuild.prefix)
+        end
+
         # Absolute path to the log directory for this package. See #logdir=
         def logdir
             if @logdir
@@ -115,17 +128,19 @@ module Autobuild
         # Returns true if this package has already been updated. It will not be
         # true if the importer has been called while Autobuild.do_update was
         # false.
-        def updated?; !!@updated end
+        def updated?
+            @updated
+        end
 
         def initialize(spec = Hash.new)
             @srcdir = @importdir = @logdir = @prefix = nil
             @updated = false
             @update = nil
             @failed = nil
-            @dependencies   = Array.new
-            @provides       = Array.new
+            @dependencies = Array.new
+            @provides = Array.new
+            @statistics = Hash.new
             @parallel_build_level = nil
-            @statistics     = Hash.new
             @failures = Array.new
             @post_install_blocks = Array.new
             @applied_post_install = false
@@ -139,19 +154,23 @@ module Autobuild
             if Hash === spec
                 name, depends = spec.to_a.first
             else
-                name, depends = spec, nil
+                name = spec
+                depends = nil
             end
 
             name = name.to_s
             @name = name
-            raise ConfigException, "package #{name} is already defined" if Autobuild::Package[name]
+            if Autobuild::Package[name]
+                raise ConfigException, "package #{name} is already defined"
+            end
+
             @@packages[name] = self
 
             # Call the config block (if any)
             yield(self) if block_given?
 
-            self.doc_utility.source_dir ||= 'doc'
-            self.doc_utility.target_dir ||= name
+            doc_utility.source_dir ||= 'doc'
+            doc_utility.target_dir ||= name
 
             # Define the default tasks
             task "#{name}-import" do
@@ -174,12 +193,10 @@ module Autobuild
                 Rake::Task["#{name}-import"].invoke
                 Rake::Task["#{name}-prepare"].invoke
                 Rake::Task["#{name}-build"].invoke
-                if has_doc? && Autobuild.do_doc
-                    Rake::Task["#{name}-doc"].invoke
-                end
+                Rake::Task["#{name}-doc"].invoke if has_doc? && Autobuild.do_doc
             end
             task :default => name
-            
+
             # The dependencies will be declared in the import phase,  so save
             # them there for now
             @spec_dependencies = depends
@@ -217,7 +234,10 @@ module Autobuild
         def to_s
             "#<#{self.class} name=#{name}>"
         end
-        def inspect; to_s end
+
+        def inspect
+            to_s
+        end
 
         # @api private
         #
@@ -226,8 +246,8 @@ module Autobuild
         #
         # @param [EnvOp] op
         # @return [void]
-        def add_env_op(op)
-            env << op
+        def add_env_op(envop)
+            env << envop
         end
 
         # Add value(s) to a list-based environment variable
@@ -300,11 +320,15 @@ module Autobuild
         def apply_env(env, set = Hash.new, ops = Array.new)
             self.env.each do |env_op|
                 next if ops.last == env_op
+
                 if env_op.type == :set
-                    if last = set[env_op.name]
+                    if (last = set[env_op.name])
                         last_pkg, last_values = *last
                         if last_values != env_op.values
-                            raise IncompatibleEnvironment, "trying to reset #{env_op.name} to #{env_op.values} in #{self.name} but this conflicts with #{last_pkg.name} already setting it to #{last_values}"
+                            raise IncompatibleEnvironment, "trying to reset "\
+                                "#{env_op.name} to #{env_op.values} in #{name} "\
+                                "but this conflicts with #{last_pkg.name} "\
+                                "already setting it to #{last_values}"
                         end
                     else
                         set[env_op.name] = [self, env_op.values]
@@ -356,9 +380,7 @@ module Autobuild
         # target files so that all the build phases of this package gets
         # retriggered. However, it should not clean the build products.
         def prepare_for_forced_build
-            if File.exist?(installstamp)
-                FileUtils.rm_f installstamp
-            end
+            FileUtils.rm_f installstamp if File.exist?(installstamp)
         end
 
         # Called when the user asked for a full rebuild. It should delete the
@@ -366,9 +388,7 @@ module Autobuild
         def prepare_for_rebuild
             prepare_for_forced_build
 
-            if File.exist?(installstamp)
-                FileUtils.rm_f installstamp
-            end
+            FileUtils.rm_f installstamp if File.exist?(installstamp)
         end
 
         # Returns true if one of the operations applied on this package failed
@@ -388,18 +408,18 @@ module Autobuild
         # will subsequently be a noop. I.e. if +build+ fails, +install+ will do
         # nothing.
         def isolate_errors(options = Hash.new)
-            if !options.kind_of?(Hash)
-                options = Hash[mark_as_failed: true]
-            end
+            options = Hash[mark_as_failed: true] unless options.kind_of?(Hash)
             options = validate_options options,
                 mark_as_failed: true,
                 ignore_errors: Autobuild.ignore_errors
 
             # Don't do anything if we already have failed
             if failed?
-                if !options[:ignore_errors]
-                    raise AlreadyFailedError, "attempting to do an operation on a failed package"
+                unless options[:ignore_errors]
+                    raise AlreadyFailedError, "attempting to do an operation "\
+                        "on a failed package"
                 end
+
                 return
             end
 
@@ -413,18 +433,12 @@ module Autobuild
                 raise
             rescue ::Exception => e
                 @failures << e
-                if options[:mark_as_failed]
-                    @failed = true
-                end
+                @failed = true if options[:mark_as_failed]
 
                 if options[:ignore_errors]
                     lines = e.to_s.split("\n")
-                    if lines.empty?
-                        lines = e.message.split("\n")
-                    end
-                    if lines.empty?
-                        lines = ["unknown error"]
-                    end
+                    lines = e.message.split("\n") if lines.empty?
+                    lines = ["unknown error"] if lines.empty?
                     message(lines.shift, :red, :bold)
                     lines.each do |line|
                         message(line)
@@ -434,9 +448,7 @@ module Autobuild
                     raise
                 end
             ensure
-                if toplevel
-                    Thread.current[:isolate_errors] = false
-                end
+                Thread.current[:isolate_errors] = false if toplevel
             end
         end
 
@@ -445,9 +457,7 @@ module Autobuild
         #
         # (see Importer#import)
         def import(options = Hash.new)
-            if !options.respond_to?(:to_hash)
-                options = Hash[only_local: options]
-            end
+            options = Hash[only_local: options] unless options.respond_to?(:to_hash)
 
             if @importer
                 result = @importer.import(self, options)
@@ -478,7 +488,8 @@ module Autobuild
         end
 
         def process_formatting_string(msg, *prefix_style)
-            prefix, suffix = [], []
+            prefix = []
+            suffix = []
             msg.split(" ").each do |token|
                 if token =~ /%s/
                     suffix << token.gsub(/%s/, name)
@@ -492,7 +503,8 @@ module Autobuild
             elsif prefix_style.empty?
                 return (prefix + suffix).join(" ")
             else
-                return [Autobuild.color(prefix.join(" "), *prefix_style), *suffix].join(" ")
+                colorized_prefix = Autobuild.color(prefix.join(" "), *prefix_style)
+                return [colorized_prefix, *suffix].join(" ")
             end
         end
 
@@ -511,17 +523,13 @@ module Autobuild
         # Display a progress message. %s in the string is replaced by the
         # package name
         def message(*args)
-            if !args.empty?
-                args[0] = "  #{process_formatting_string(args[0])}"
-            end
+            args[0] = "  #{process_formatting_string(args[0])}" unless args.empty?
             Autobuild.message(*args)
         end
 
         def progress_start(*args, done_message: nil, **raw_options, &block)
             args[0] = process_formatting_string(args[0], :bold)
-            if done_message
-                done_message = process_formatting_string(done_message)
-            end
+            done_message = process_formatting_string(done_message) if done_message
             Autobuild.progress_start(self, *args,
                 done_message: done_message, **raw_options, &block)
         end
@@ -559,13 +567,14 @@ module Autobuild
         end
 
         def run(*args, &block)
-            if args.last.kind_of?(Hash)
-                options = args.pop
-            else
-                options = Hash.new
-            end
+            options =
+                if args.last.kind_of?(Hash)
+                    args.pop
+                else
+                    Hash.new
+                end
             options[:env] = options.delete(:resolved_env) ||
-                (options[:env] || Hash.new).merge(resolved_env)
+                            (options[:env] || Hash.new).merge(resolved_env)
             Autobuild::Subprocess.run(self, *args, options, &block)
         end
 
@@ -596,17 +605,49 @@ module Autobuild
             task
         end
 
-        def doc_dir=(value); doc_utility.source_dir = value end
-        def doc_dir; doc_utility.source_dir end
-        def doc_target_dir=(value); doc_utility.target_dir = value end
-        def doc_target_dir; doc_utility.target_dir end
-        def doc_task(&block); doc_utility.task(&block) end
-        def generates_doc?; doc_utility.enabled? end
-        def enable_doc; doc_utility.enabled = true end
-        def disable_doc; doc_utility.enabled = false end
-        def install_doc; doc_utility.install end
-        def doc_disabled; doc_utility.disabled end
-        def has_doc?; doc_utility.has_task? end
+        def doc_dir=(value)
+            doc_utility.source_dir = value
+        end
+
+        def doc_dir
+            doc_utility.source_dir
+        end
+
+        def doc_target_dir=(value)
+            doc_utility.target_dir = value
+        end
+
+        def doc_target_dir
+            doc_utility.target_dir
+        end
+
+        def doc_task(&block)
+            doc_utility.task(&block)
+        end
+
+        def generates_doc?
+            doc_utility.enabled?
+        end
+
+        def enable_doc
+            doc_utility.enabled = true
+        end
+
+        def disable_doc
+            doc_utility.enabled = false
+        end
+
+        def install_doc
+            doc_utility.install
+        end
+
+        def doc_disabled
+            doc_utility.disabled
+        end
+
+        def has_doc?
+            doc_utility.has_task?
+        end
 
         def post_install(*args, &block)
             if args.empty?
@@ -646,7 +687,7 @@ module Autobuild
         def all_dependencies(result = Set.new)
             dependencies.each do |pkg_name|
                 pkg = Autobuild::Package[pkg_name]
-                if !result.include?(pkg.name)
+                unless result.include?(pkg.name)
                     result << pkg.name
                     pkg.all_dependencies(result)
                 end
@@ -666,18 +707,21 @@ module Autobuild
         def depends_on(*packages)
             packages.each do |p|
                 p = p.name if p.respond_to?(:name)
-                raise ArgumentError, "#{p.inspect} should be a string" if !p.respond_to? :to_str
+                unless p.respond_to?(:to_str)
+                    raise ArgumentError, "#{p.inspect} should be a string"
+                end
+
                 p = p.to_str
                 next if p == name
-                unless pkg = Package[p]
-                    raise ConfigException.new(self), "package #{p}, listed as a dependency of #{self.name}, is not defined"
+
+                unless (pkg = Package[p])
+                    raise ConfigException.new(self), "package #{p}, "\
+                        "listed as a dependency of #{name}, is not defined"
                 end
 
                 next if @dependencies.include?(pkg.name)
 
-                if Autobuild.verbose
-                    Autobuild.message "#{name} depends on #{pkg.name}"
-                end
+                Autobuild.message "#{name} depends on #{pkg.name}" if Autobuild.verbose
 
                 task "#{name}-import"  => "#{pkg.name}-import"
                 task "#{name}-prepare" => "#{pkg.name}-prepare"
@@ -690,16 +734,17 @@ module Autobuild
         # listed in +packages+ are aliases for this package.
         def provides(*packages)
             packages.each do |p|
-                raise ArgumentError, "#{p.inspect} should be a string" if !p.respond_to? :to_str
+                unless p.respond_to?(:to_str)
+                    raise ArgumentError, "#{p.inspect} should be a string"
+                end
+
                 p = p.to_str
                 next if p == name
                 next if @provides.include?(name)
 
-                @@provides[p] = self 
+                @@provides[p] = self
 
-                if Autobuild.verbose
-                    Autobuild.message "#{name} provides #{p}"
-                end
+                Autobuild.message "#{name} provides #{p}" if Autobuild.verbose
 
                 task p => name
                 task "#{p}-import" => "#{name}-import"
@@ -712,13 +757,11 @@ module Autobuild
         # Iterates on all available packages
         # if with_provides is true, includes the list
         # of package aliases
-        def self.each(with_provides = false, &p)
-            if !p
-                return enum_for(:each, with_provides)
-            end
+        def self.each(with_provides = false, &block)
+            return enum_for(:each, with_provides) unless block
 
-            @@packages.each(&p) 
-            @@provides.each(&p) if with_provides
+            @@packages.each(&block)
+            @@provides.each(&block) if with_provides
         end
 
         # Gets a package from its name
@@ -743,7 +786,7 @@ module Autobuild
         end
 
         # Returns the level of parallelism authorized during the build for this
-        # particular package. If not set, defaults to the system-wide option 
+        # particular package. If not set, defaults to the system-wide option
         # (Autobuild.parallel_build_level and Autobuild.parallel_build_level=).
         #
         # The default value is the number of CPUs on this system.
@@ -764,7 +807,6 @@ module Autobuild
         def in_dir(directory)
             @in_dir_stack << directory
             yield
-
         ensure
             @in_dir_stack.pop
         end
@@ -792,21 +834,26 @@ module Autobuild
         end
 
         def utility(utility_name)
-            utilities[utility_name] ||= Autobuild.create_utility(utility_name, self)
+            utilities[utility_name.to_s] ||= Autobuild.create_utility(utility_name, self)
         end
 
+        def respond_to_missing?(name, _include_all)
+            utilities.key?(name.to_s)
+        end
 
-        def method_missing(m, *args, &block)
-            case m.to_s
+        def method_missing(name, *args, &block)
+            case name.to_s
             when /(\w+)_utility$/
                 utility_name = $1
-                if !args.empty?
+
+                unless args.empty?
                     raise ArgumentError, "expected 0 arguments and got #{args.size}"
                 end
+
                 begin
                     return utility(utility_name)
                 rescue ArgumentError => e
-                    raise NoMethodError.new(m), e.message, e.backtrace
+                    raise NoMethodError.new(name), e.message, e.backtrace
                 end
             end
             super
@@ -821,4 +868,3 @@ module Autobuild
         end
     end
 end
-
